BINIUS-247: deduplicate the circuit interpreters via a shared execution-context trait
The single-instance and batched circuit interpreters ran the same bytecode
through the same ~40-opcode dispatch, every operation handler, and the same
bytecode readers. All of it was written twice, differing only in where a
decoded instruction reads and writes: one value vector versus one column per
instance. Changing an opcode meant editing both copies identically or the two
provers would silently disagree.

Name the one thing that differs behind a small trait over the execution
context: how many instances, read register r of instance i, write register r
of instance i, note an assertion failure for instance i. Write the opcode
logic once, as "for each instance, apply the operation". The single-instance
interpreter becomes the degenerate case of one instance; the batched one is the
many-column case. Both interpreters are now thin wrappers over the shared
executor.

Pure refactor, no behavior change: the assertion messages and blamed-instance
index are byte-identical, and the batched-equals-single equivalence test is
untouched and still passes.
